fix: moved services to common packages
- moved axios, becknApi, api service to common
- modified respective files in retail app
- integrated common api services to retails app
