+++ apps/retail/pages/checkout.tsx
@@ -4,7 +4,6 @@ import { Box, useToast, useTheme } from '@chakra-ui/react'
 import { useLanguage } from '../hooks/useLanguage'
 
 import { CustomToast } from '@components/signIn/SignIn'
-import { useInitMutation } from '@services/init'
 
 import { Checkout } from '@beckn-ui/becknified-components'
 import { FormData, FormField } from '@beckn-ui/molecules'
@@ -17,6 +16,7 @@ import { DiscoveryRootState, ICartRootState, PaymentBreakDownModel, QuoteBreakup
 import { checkoutActions, CheckoutRootState } from '@beckn-ui/common/src/store/checkout-slice'
 import { cartActions } from '@beckn-ui/common/src/store/cart-slice'
 import { DOMAIN } from '@beckn-ui/common'
+import { useInitMutation } from '@beckn-ui/common/src/services/init'
 
 export type ShippingFormData = {
   name: string
