+++ apps/retail/pages/signUp.tsx
@@ -5,10 +5,10 @@ import { signUpValidateForm } from '@utils/form-utils'
 import { BecknAuth } from '@beckn-ui/becknified-components'
 import Router from 'next/router'
 import { Box, useBreakpoint } from '@chakra-ui/react'
-import { useRegisterMutation } from '@services/Users'
 import Logo from '@public/images/Logo.svg'
 import { FetchBaseQueryError } from '@reduxjs/toolkit/query'
 import { FormErrors, SignInResponse, SignUpProps } from '@beckn-ui/common/lib/types'
+import { useRegisterMutation } from '@beckn-ui/common/src/services/User'
 
 const SignUp = () => {
   const { t } = useLanguage()
