--- apps/retail/services/becknApi.ts
@@ -1,59 +1,0 @@
-import { createApi, fetchBaseQuery, retry } from '@reduxjs/toolkit/query/react'
-import type { BaseQueryFn, FetchArgs, FetchBaseQueryError } from '@reduxjs/toolkit/query'
-import { feedbackActions } from '@beckn-ui/common/src/store/ui-feedback-slice'
-
-// Create our baseQuery instance
-const baseQuery = fetchBaseQuery({
-  baseUrl: process.env.NEXT_PUBLIC_API_URL
-})
-
-const baseQueryWithRetry = retry(baseQuery, { maxRetries: 6 })
-
-const baseQueryWithErrorHandling: BaseQueryFn<string | FetchArgs, unknown, FetchBaseQueryError> = async (
-  args,
-  api,
-  extraOptions
-) => {
-  const result = await baseQueryWithRetry(args, api, extraOptions)
-  if (result.error && result.error.status === 400) {
-    const { message } = (result.error.data as any)?.error || 'Something went wrong'
-    api.dispatch(
-      feedbackActions.setToastData({
-        toastData: { message: 'Error!', display: true, type: 'error', description: message }
-      })
-    )
-  }
-  return result
-}
-
-/**
- * Create a base API to inject endpoints into elsewhere.
- * Components using this API should import from the injected site,
- * in order to get the appropriate types,
- * and to ensure that the file injecting the endpoints is loaded
- */
-export const api = createApi({
-  /**
-   * `reducerPath` is optional and will not be required by most users.
-   * This is useful if you have multiple API definitions,
-   * e.g. where each has a different domain, with no interaction between endpoints.
-   * Otherwise, a single API definition should be used in order to support tag invalidation,
-   * among other features
-   */
-  reducerPath: 'splitApi',
-  /**
-   * A bare bones base query would just be `baseQuery: fetchBaseQuery({ baseUrl: '/' })`
-   */
-  baseQuery: baseQueryWithErrorHandling,
-  /**
-   * Tag types must be defined in the original API definition
-   * for any tags that would be provided by injected endpoints
-   */
-  tagTypes: ['Search', 'Select'],
-  /**
-   * This api has endpoints injected in adjacent files,
-   * which is why no endpoints are shown below.
-   * If you want all endpoints defined in the same file, they could be included here instead
-   */
-  endpoints: () => ({})
-})
