+++ packages/common/src/services/api.ts
@@ -40,7 +40,7 @@ const baseQueryWithReauth: BaseQueryFn<string | FetchArgs, unknown, FetchBaseQue
  * in order to get the appropriate types,
  * and to ensure that the file injecting the endpoints is loaded
  */
-export const api = createApi({
+const api = createApi({
   /**
    * `reducerPath` is optional and will not be required by most users.
    * This is useful if you have multiple API definitions,
@@ -65,3 +65,5 @@ export const api = createApi({
    */
   endpoints: () => ({})
 })
+
+export default api
