+++ packages/common/src/services/becknApi.ts
@@ -1,24 +1,15 @@
 import { createApi, fetchBaseQuery, retry } from '@reduxjs/toolkit/query/react'
 import type { BaseQueryFn, FetchArgs, FetchBaseQueryError } from '@reduxjs/toolkit/query'
-import { RootState } from '../store'
 import { feedbackActions } from '@beckn-ui/common/src/store/ui-feedback-slice'
 
 // Create our baseQuery instance
 const baseQuery = fetchBaseQuery({
-  baseUrl: process.env.NEXT_PUBLIC_STRAPI_URL,
-  prepareHeaders: (headers, { getState }) => {
-    // By default, if we have a token in the store, let's use that for authenticated requests
-    const token = (getState() as RootState).auth.jwt
-    if (token) {
-      headers.set('authentication', `Bearer ${token}`)
-    }
-    return headers
-  }
+  baseUrl: process.env.NEXT_PUBLIC_API_URL
 })
 
-const baseQueryWithRetry = retry(baseQuery, { maxRetries: 0 })
+const baseQueryWithRetry = retry(baseQuery, { maxRetries: 6 })
 
-const baseQueryWithReauth: BaseQueryFn<string | FetchArgs, unknown, FetchBaseQueryError> = async (
+const baseQueryWithErrorHandling: BaseQueryFn<string | FetchArgs, unknown, FetchBaseQueryError> = async (
   args,
   api,
   extraOptions
@@ -41,7 +32,7 @@ const baseQueryWithReauth: BaseQueryFn<string | FetchArgs, unknown, FetchBaseQue
  * in order to get the appropriate types,
  * and to ensure that the file injecting the endpoints is loaded
  */
-export const api = createApi({
+const api = createApi({
   /**
    * `reducerPath` is optional and will not be required by most users.
    * This is useful if you have multiple API definitions,
@@ -53,12 +44,12 @@ export const api = createApi({
   /**
    * A bare bones base query would just be `baseQuery: fetchBaseQuery({ baseUrl: '/' })`
    */
-  baseQuery: baseQueryWithReauth,
+  baseQuery: baseQueryWithErrorHandling,
   /**
    * Tag types must be defined in the original API definition
    * for any tags that would be provided by injected endpoints
    */
-  tagTypes: ['Auth'],
+  tagTypes: ['Search', 'Select'],
   /**
    * This api has endpoints injected in adjacent files,
    * which is why no endpoints are shown below.
@@ -66,3 +57,5 @@ export const api = createApi({
    */
   endpoints: () => ({})
 })
+
+export default api
